+++ apps/dashboard/src/app/layout.tsx
@@ -19,6 +19,7 @@ import './globals.css';
 import { CSPostHogProvider, UserIdentity } from './providers';
 import dynamic from 'next/dynamic';
 import { SpeedInsights } from "@vercel/speed-insights/next";
+import { VersionAlerter } from '../components/version-alerter';
 
 export const metadata: Metadata = {
   metadataBase: new URL(process.env.NEXT_PUBLIC_STACK_URL || ''),
@@ -98,6 +99,7 @@ export default function RootLayout({
                 <ClientPolyfill />
                 <RouterProvider>
                   <UserIdentity />
+                  <VersionAlerter severeOnly={false} />
                   {children}
                 </RouterProvider>
               </StackTheme>