+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack-sc
 
+## 2.6.8
+
+## 2.6.7
+
+### Patch Changes
+
+- Bugfixes
+
 ## 2.6.6
 
 ## 2.6.5