+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-shared
 
+## 2.6.8
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-sc@2.6.8
+
+## 2.6.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-sc@2.6.7
+
 ## 2.6.6
 
 ### Patch Changes